requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
asgiref==3.8.1
assemblyai==0.20.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.1.1
deepeval==1.1.6
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.9
distro==1.9.0
Django==5.1.6
django-admin==2.0.2
django-cors-headers==3.14.0
django-excel-response2==3.0.5
django-model-utils==4.3.1
django-redis==5.4.0
django-six==1.0.5
dnspython==2.6.1
docx2txt==0.8
elevenlabs==1.50.6
email_validator==2.2.0
excel-base==1.0.4
exceptiongroup==1.2.0
execnet==2.1.1
executing==1.2.0
Faker==36.1.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.3.3
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.11.0
google-api-python-client==2.143.0
google-auth==2.17.3
google-auth-httplib2==0.2.0
google-cloud-dialogflow==2.21.0
google-cloud-dialogflow-cx==1.23.0
google-generativeai==0.7.2
googleapis-common-protos==1.59.0
greenlet==3.0.3
grpcio==1.63.2
grpcio-status==1.53.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.4
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.11.0
isoweek==1.3.3
jedi==0.18.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.2.10
langchain-chroma==0.1.2
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.9
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==4.22.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyactiveresource==2.2.1
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.14.0
pylance==0.21.0
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
pytest==8.3.2
pytest-django==4.9.0
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
ragas==0.1.16
redis==5.2.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
screen==1.0.1
sentry-sdk==2.13.0
shellingham==1.5.4
ShopifyAPI==8.0.1
six==1.15.0
sniffio==1.3.0
SQLAlchemy==2.0.31
sqlparse==0.3.1
stack-data==0.6.2
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tenacity==8.4.2
tiktoken==0.7.0
TimeConvert==3.0.13
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.5
traitlets==5.9.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.30.3
uvloop==0.19.0
uWSGI==2.0.27
virtualenv==20.29.2
watchfiles==0.22.0
wcwidth==0.2.6
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xlwt==1.3.0
xxhash==3.5.0
yarl==1.8.2
zipp==3.19.2
